add typecode cache via a mixin class
